run(), each(), close() をファイル間で共通化
