+++ 05.asynchronous_sqlite3/async_await_with_error.js
@@ -1,34 +1,7 @@
 #!/usr/bin/env node
 
 import sqlite3 from "sqlite3";
-
-function run(db, sql, ...params) {
-  return new Promise((resolve) => {
-    db.run(sql, ...params, function () {
-      resolve(this);
-    });
-  });
-}
-
-function each(db, sql, callback) {
-  return new Promise((resolve) => {
-    db.each(
-      sql,
-      (err, row) => {
-        callback(row);
-      },
-      () => {
-        resolve();
-      }
-    );
-  });
-}
-
-function close(db) {
-  return new Promise((resolve) => {
-    db.close(() => resolve());
-  });
-}
+import { run, each, close } from "./functions.js";
 
 async function main() {
   const db = new sqlite3.Database(":memory:");
